fix: improve behaviour/fix bugs
This library had some unaddressed buggy behaviour, so made some small
changes that improve UX on both iOS and Android.

One change required doing some additional math solely needed
on Android; unclear why, but it seemed to solve the issue where it was
very hard to drag a photo into the 1st tile position. Another change was
introducing new onPressIn and onPressOut event handlers which work
better for disabling/enabling scrolling on the continer. Finally, allowed
the consumer to control the delayLongPress parameter to reduce the
long-press timeout to initiate the drag-n-drag, aligning behaviour more
closely to the original, unforked version of this library.